Don't periodically chown volumes for user namespaces
This periodic chown makes everything in shared volumes
owned by root. This is very frustrating for log volumes
where a file may be created as say `www-data`, but then
chowned by this sync process later to root.

The downside is that, with user namespaces, volumes are created
with very lax permissions:

    drwxrwxrwx    2 nobody  nogroup   68 Feb  9 22:00 logs

Maybe should do a one time chown instead of a perodic one?
Or maybe it is just left as an exercise to the pod owner to
set their own permissions?
(I think I vote the latter, because it is nice to be able to run containers
as non-root, and with permissions like that, they will actually
be able to use volumes)